+++ changelog.md
@@ -2,12 +2,14 @@
 
     - Youtube URLs can be played piping the audio using `youtube-dl`.
     - New method for discovering local IP in macOS. This will improve the
-      stability of the system tray application. See
+      stability of the system tray application. Closes
       [#5](https://github.com/muammar/mkchromecast/issues/5).
     - Now it is possible to select a different color for the system tray icon.
       This is useful when using dark themes.
     - The node streaming has been updated to work with `node v6.3.0`.
     - Improved stability when using system tray icon.
+    - New method in preferences window to reset to default configurations.
+      Closes [#6](https://github.com/muammar/mkchromecast/issues/6).
 
 * mkchromecast (0.2.9.1) **released**: 2016/06/29